# openapi/vehiclecare-swagger.yml
openapi: 3.0.3
info:
  title: VehicleCare API
  version: 1.0.0
  description: >
    Contrato OpenAPI del backend `vehiclecare-micro` basado en los endpoints
    actualmente usados por la app y expuestos por los controladores activos.
servers:
  - url: http://localhost:8080
    description: Entorno local por defecto
...
paths:
...
  /users/{id}:
    get:
      tags: [Profile]
      summary: Obtener perfil del usuario autenticado
      operationId: getUserProfile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Perfil encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Usuario no encontrado
  /users/{id}/profile:
    put:
      tags: [Profile]
      summary: Actualizar nombre del perfil
      operationId: updateUserProfile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
            example:
              name: Alvaro
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}/profile-image:
    post:
      tags: [Profile]
      summary: Subir imagen de perfil
      operationId: uploadUserProfileImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SingleFileUploadRequest'
      responses:
        '200':
          description: Imagen de perfil actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserIdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Identificador del usuario
...
  responses:
    BadRequest:
      description: Solicitud inválida o error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            timestamp: '2026-05-10T10:00:00Z'
            status: 400
            error: Validación incorrecta
            path: /vehicles
            details:
              - field: licensePlate
                message: La matrícula debe tener formato español 1234ABC o 1234 ABC
    Unauthorized:
      description: No autenticado o token inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            timestamp: '2026-05-10T10:00:00Z'
            status: 401
            error: Unauthorized
            path: /vehicles
            details:
              - field: authorization
                message: Falta el token Bearer
    Forbidden:
      description: Acceso denegado por validación de ownership
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            timestamp: '2026-05-10T10:00:00Z'
            status: 403
            error: Acceso denegado
            path: /users/other-user-id
            details:
              - field: general
                message: No puedes acceder al perfil de otro usuario
...
  schemas:
...
    UserProfileUpdateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 10
          pattern: ^[A-Za-zÁÉÍÓÚáéíóúÑñ ]+$
    UserResponse:
      type: object
      required: [id, name, email, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        profileImageUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SingleFileUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
...
    ApiError:
      type: object
      required: [timestamp, status, error, path, details]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        path:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiFieldError'
    ApiFieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string